[feature/matrix_func_refactor] move matrix calculation from Gower class to a util function
