+++ pyproject.toml
@@ -10,6 +10,7 @@ dependencies = [
     "numpy>=2.2.6",
     "pandas>=2.3.0",
     "pydantic>=2.12.5",
+    "rpy2>=3.6.7",
     "scikit-learn>=1.7.2",
     "scipy>=1.15.3",
     "tqdm>=4.67.1",
