+++ tests/gower_api/unit/test_matrix_coverage.py
@@ -6,6 +6,7 @@ import pytest
 
 from gower_metric import Config, Gower
 from gower_metric.utils.matrix.convert_matrix import get_scipy_sparse_matrix
+from gower_metric.utils.matrix.distance import calculate_matrix
 
 
 class TestSparseMatrixConversion:
@@ -28,24 +29,25 @@ class TestMatrixCompute:
 
     def test_similarity_matrix_diagonal_ones(self) -> None:
         gower, data = self._get_fitted_gower()
-        mat = gower.matrix(data, matrix_type="similarity")
+        mat = calculate_matrix(gower, data, matrix_type="similarity")
         np.testing.assert_array_almost_equal(np.diag(mat), [1.0, 1.0, 1.0])
         assert mat[0, 1] == mat[1, 0]  # symmetric
 
     def test_distance_matrix_diagonal_zeros(self) -> None:
         gower, data = self._get_fitted_gower()
-        mat = gower.matrix(data, matrix_type="distance")
+        mat = calculate_matrix(gower, data, matrix_type="distance")
         np.testing.assert_array_almost_equal(np.diag(mat), [0.0, 0.0, 0.0])
 
     def test_verbose_matrix(self) -> None:
         gower, data = self._get_fitted_gower()
-        mat = gower.matrix(data, verbose=1)
+        mat = calculate_matrix(gower, data, verbose=1)
         assert mat.shape == (3, 3)
 
     def test_sparse_similarity_matrix(self) -> None:
         """Similarity + convert_to_sparse covers branch 164→167."""
         gower, data = self._get_fitted_gower()
-        gower.matrix(
+        calculate_matrix(
+            gower,
             data,
             matrix_type="similarity",
             convert_to_sparse=True,
@@ -55,12 +57,12 @@ class TestMatrixCompute:
     def test_similarity_matrix_pandas(self) -> None:
         """Matrix computation with pandas DataFrame."""
         gower, data = self._get_fitted_gower_pandas()
-        mat = gower.matrix(data, matrix_type="similarity")
+        mat = calculate_matrix(gower, data, matrix_type="similarity")
         np.testing.assert_array_almost_equal(np.diag(mat), [1.0, 1.0, 1.0])
 
     def test_distance_matrix_pandas(self) -> None:
         gower, data = self._get_fitted_gower_pandas()
-        mat = gower.matrix(data, matrix_type="distance")
+        mat = calculate_matrix(gower, data, matrix_type="distance")
         np.testing.assert_array_almost_equal(np.diag(mat), [0.0, 0.0, 0.0])
 
     def test_matrix_float64_dtype(self) -> None:
@@ -68,7 +70,7 @@ class TestMatrixCompute:
         data = np.array([[1.0], [5.0], [10.0]])
         cfg = Config(feature_types={0: "numeric"}, data_type=np.float64)
         gower = Gower(cfg).fit(data)
-        mat = gower.matrix(data)
+        mat = calculate_matrix(gower, data)
         assert mat.dtype == np.float64
         assert mat.shape == (3, 3)
 
@@ -77,7 +79,7 @@ class TestMatrixCompute:
         data = pd.DataFrame({"x": [1.0, 5.0, 10.0]})
         cfg = Config(feature_types={"x": "numeric"}, data_type=np.float64)
         gower = Gower(cfg).fit(data)
-        mat = gower.matrix(data)
+        mat = calculate_matrix(gower, data)
         assert mat.dtype == np.float64
